+++ akka-stream/src/main/scala/akka/stream/impl/Messages.scala
@@ -3,8 +3,6 @@
  */
 package akka.stream.impl
 
-import org.reactivestreams.Subscription
-
 /**
  * INTERNAL API
  */
@@ -12,7 +10,7 @@ private[akka] case object SubscribePending
 /**
  * INTERNAL API
  */
-private[akka] case class RequestMore(subscription: ActorSubscription[_], demand: Int)
+private[akka] case class RequestMore(subscription: ActorSubscription[_], demand: Long)
 /**
  * INTERNAL API
  */
