+++ akka-stream/src/test/scala/akka/stream/FlowDropWithinSpec.scala
@@ -18,19 +18,19 @@ class FlowDropWithinSpec extends AkkaSpec {
 
     "deliver elements after the duration, but not before" in {
       val input = Iterator.from(1)
-      val p = StreamTestKit.PublisherProbe[Int]()
-      val c = StreamTestKit.SubscriberProbe[Int]()
+      val p = StreamTestKit.PublisherProbe[Long]()
+      val c = StreamTestKit.SubscriberProbe[Long]()
       Flow(p).dropWithin(1.second).produceTo(materializer, c)
       val pSub = p.expectSubscription
       val cSub = c.expectSubscription
       cSub.request(100)
       val demand1 = pSub.expectRequest
-      (1 to demand1) foreach { _ ⇒ pSub.sendNext(input.next()) }
+      (1L to demand1) foreach { _ ⇒ pSub.sendNext(input.next()) }
       val demand2 = pSub.expectRequest
-      (1 to demand2) foreach { _ ⇒ pSub.sendNext(input.next()) }
+      (1L to demand2) foreach { _ ⇒ pSub.sendNext(input.next()) }
       val demand3 = pSub.expectRequest
       c.expectNoMsg(1500.millis)
-      (1 to demand3) foreach { _ ⇒ pSub.sendNext(input.next()) }
+      (1L to demand3) foreach { _ ⇒ pSub.sendNext(input.next()) }
       ((demand1 + demand2 + 1) to (demand1 + demand2 + demand3)) foreach { n ⇒ c.expectNext(n) }
       pSub.sendComplete()
       c.expectComplete
